Remove utils class
