--- spec/unit/opencensus/trace/exporters/jaeger_driver_spec.rb
@@ -1,6 +1,0 @@
-require 'spec_helper'
-
-describe OpenCensus::Trace::Exporters::JaegerDriver do
-  let(:service_name) { 'service-name' }
-  let(:tags) { {} }
-end
